INSP: move checks for anchored paths to separate inspection
