+++ src/main/resources/inspectionDescriptions/RsAnchoredPaths.html
@@ -1,0 +1,8 @@
+<html>
+<body>
+Checks that all paths in <code>use</code> items start with a crate name, or with <code>crate</code>,
+<code>super</code>, or <code>self</code>.<br>
+
+Related to "anchored paths" variant in Edition 2018.
+</body>
+</html>
